Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary>=2.9.7
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.6
pydantic[email]==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
python-dotenv==1.0.0
